pyproject.toml
[project]
name = "incident-sim"
version = "0.1.0"
description = "ITIL Incident Simulation Training Environment"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.49.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "sqlmodel>=0.0.22",
    "textual>=0.87.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "eval_type_backport>=0.1.3",
]

[project.scripts]
incident-sim-api = "backend.main:main"
incident-sim-tui = "tui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "tui"]
